backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
langchain-google-genai
langchain-community
langgraph
supabase
pypdf
pydantic